[UI] Consolidate legacy themes/ directory into theme/ module
Complete the planned Phase 1 consolidation of ui/themes into ui/theme:

- Move themes/hooks.tsx to theme/hooks.tsx; drop the redundant nested
  ProviderStoreWrapper and replace the hand-rolled matchMedia hook with
  Sistent's useMediaQuery (defaultMatches: true preserves the dark-first
  initial render and adds live OS theme tracking).
- Move themes/rjsf.ts to theme/rjsf.ts; remove the throwaway createTheme()
  used only for default breakpoints (callback styleOverrides instead) and
  the duplicate default export.
- Split themes/App.styles.tsx: app-shell layout styles to
  components/App.styles.tsx, ThemeResponsiveSnackbar to theme/snackbar.tsx
  (its documented target home), hoisting static style objects.
- Delete themes/app.ts and themes/index.ts: the six live hex values are
  byte-identical to notificationColors exported by @sistent/sistent, so
  NotificationCenter constants and ConnectionChip now read the design
  system constant; dead exports (Colors, darkNotificationColors,
  SUCCESS_V2) removed.
- Remove @/themes aliases from tsconfig/jsconfig/vitest, the @/themes*
  lint restrictions, six now-unneeded defensive test mocks, and stale
  carve-outs (constants/** hex exemption, nonexistent UserPreferences
  allowlist path, dead THEMING.md migration rows).
- Update THEMING.md, ARCHITECTURE.md, and the restructure plan checklist.
