+++ ui/components/layout/NotificationCenter/constants.tsx
@@ -1,8 +1,7 @@
-import { NOTIFICATIONCOLORS } from '../../../themes';
 import AlertIcon from '../../../assets/icons/AlertIcon';
 import ErrorIcon from '../../../assets/icons/ErrorIcon';
 import ReadIcon from '../../../assets/icons/ReadIcon';
-import { InfoIcon } from '@sistent/sistent';
+import { InfoIcon, notificationColors } from '@sistent/sistent';
 import type { Theme } from '@sistent/sistent';
 
 export const SEVERITY = {
@@ -41,23 +40,23 @@ export const getStatusStyle = (theme: Theme) => ({
 export const SEVERITY_STYLE = {
   [SEVERITY.INFO]: {
     icon: InfoIcon,
-    color: NOTIFICATIONCOLORS.INFO,
-    darkColor: NOTIFICATIONCOLORS.INFO,
+    color: notificationColors.info.main,
+    darkColor: notificationColors.info.main,
   },
   [SEVERITY.ERROR]: {
     icon: ErrorIcon,
-    color: NOTIFICATIONCOLORS.ERROR,
-    darkColor: NOTIFICATIONCOLORS.ERROR_DARK,
+    color: notificationColors.error.main,
+    darkColor: notificationColors.error.dark,
   },
   [SEVERITY.WARNING]: {
     icon: AlertIcon,
-    color: NOTIFICATIONCOLORS.WARNING,
-    darkColor: NOTIFICATIONCOLORS.WARNING,
+    color: notificationColors.warning.main,
+    darkColor: notificationColors.warning.main,
   },
   [SEVERITY.SUCCESS]: {
     icon: InfoIcon,
-    color: NOTIFICATIONCOLORS.SUCCESS,
-    darkColor: NOTIFICATIONCOLORS.SUCCESS,
+    color: notificationColors.success.main,
+    darkColor: notificationColors.success.main,
   },
 };
