--- ui/themes/app.ts
@@ -1,22 +1,0 @@
-export const Colors = {
-  darkJungleGreen: '#1E2117',
-  caribbeanGreen: '#00D3a9',
-  keppelGreen: '#00B39F',
-  charcoal: '#3C494F',
-};
-
-export const notificationColors = {
-  error: '#F91313',
-  warning: '#F0A303',
-  success: '#206D24',
-  info: '#2196F3',
-  darkRed: '#B32700',
-  lightwarning: '#E75225',
-};
-
-export const darkNotificationColors = {
-  error: '#F91313',
-  warning: '#F0D053',
-  success: '#78C57C',
-  info: '#5FD4FF',
-};
