--- ui/themes/index.ts
@@ -1,13 +1,0 @@
-import { rjsfTheme } from './rjsf';
-import { notificationColors } from './app';
-
-export const NOTIFICATIONCOLORS = {
-  ERROR: '#F91313',
-  WARNING: '#F0A303',
-  SUCCESS: '#206D24',
-  INFO: '#2196F3',
-  SUCCESS_V2: '#3FC6B6',
-  ERROR_DARK: '#B32700',
-};
-
-export { rjsfTheme, notificationColors };
